refactoring: работа с репо в DetailScreenActivity
